eagle-final: EAGLE3 async spec decode with all bug fixes
- int64 fix for Triton store_kvcache_kernel (slot overflow at high block IDs)
- SHM buffer 256MB for B=16 eagle data
- bf16 dtype fixes in model_runner warmup and FlashInfer plan
- hidden_states padding in tree decode CUDA graph for B < bucket
- extend-after-decode v5c with target KV refresh
- FlashInfer custom mask glue decode integration
- CPU mask precompute optimization with stream sync fix

Tested: B=1/2/16 CG+eager, 512/512 seqs --all, 255 tok/s k=7

+++ ssd/models/eagle3_draft_llama3.py
@@ -282,7 +282,7 @@ class Eagle3DraftForCausalLM(nn.Module):
                 self._save_debug_inputs(input_ids, positions, hidden_states)
                 self._debug_saved = True
             
-            hidden_states_projected = self.fc(hidden_states)  # [num_tokens, d_model_draft]
+            hidden_states_projected = self.fc(hidden_states.to(self.fc.weight.dtype))  # [num_tokens, d_model_draft]
         else:
             hidden_states_projected = hidden_states # draft self-conditioning output, already d_model_draft from prenorm
